refactor to call vllm2
